[BE] refactor : 학과 관리 패키지 수정
