+++ backend/src/main/java/moim_today/application/admin/department/AdminDepartmentService.java
@@ -1,0 +1,13 @@
+package moim_today.application.admin.department;
+
+import moim_today.dto.department.ApproveRequestDepartmentRequest;
+import moim_today.dto.department.RequestDepartmentResponse;
+
+import java.util.List;
+
+public interface AdminDepartmentService {
+
+    List<RequestDepartmentResponse> findAll();
+
+    void approveRequest(final ApproveRequestDepartmentRequest approveRequestDepartmentRequest);
+}
