+++ backend/src/main/java/moim_today/application/department/request_department/RequestDepartmentServiceImpl.java
@@ -1,14 +1,8 @@
 package moim_today.application.department.request_department;
 
 import moim_today.dto.department.AddDepartmentRequest;
-import moim_today.dto.department.ApproveRequestDepartmentRequest;
-import moim_today.dto.department.RequestDepartmentResponse;
-import moim_today.implement.department.department.DepartmentAppender;
 import moim_today.implement.department.request_department.RequestDepartmentAppender;
-import moim_today.implement.department.request_department.RequestDepartmentFinder;
-import moim_today.implement.department.request_department.RequestDepartmentRemover;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -16,34 +10,9 @@ import java.util.List;
 public class RequestDepartmentServiceImpl implements RequestDepartmentService {
 
     private final RequestDepartmentAppender requestDepartmentAppender;
-    private final RequestDepartmentFinder requestDepartmentFinder;
-    private final RequestDepartmentRemover requestDepartmentRemover;
-    private final DepartmentAppender departmentAppender;
 
-    public RequestDepartmentServiceImpl(final RequestDepartmentAppender requestDepartmentAppender,
-                                        final RequestDepartmentFinder requestDepartmentFinder,
-                                        final RequestDepartmentRemover requestDepartmentRemover,
-                                        final DepartmentAppender departmentAppender) {
+    public RequestDepartmentServiceImpl(final RequestDepartmentAppender requestDepartmentAppender) {
         this.requestDepartmentAppender = requestDepartmentAppender;
-        this.requestDepartmentFinder = requestDepartmentFinder;
-        this.requestDepartmentRemover = requestDepartmentRemover;
-        this.departmentAppender = departmentAppender;
-    }
-
-    @Override
-    public List<RequestDepartmentResponse> findAll() {
-        return requestDepartmentFinder.findAll();
-    }
-
-    @Transactional
-    @Override
-    public void approveRequest(final ApproveRequestDepartmentRequest approveRequestDepartmentRequest) {
-        departmentAppender.addDepartment(
-                approveRequestDepartmentRequest.universityId(),
-                approveRequestDepartmentRequest.requestDepartmentName()
-        );
-
-        requestDepartmentRemover.deleteById(approveRequestDepartmentRequest.requestDepartmentId());
     }
 
     @Override
